tools/requirements.txt
# Python dependencies for thesis automation tools
# Install with: pip install -r requirements.txt

# Google Cloud APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Environment variable management
python-dotenv==1.0.0

# Data processing
pandas==2.1.3
openpyxl==3.1.2  # For Excel export

# AI APIs (optional, for qualitative analysis)
openai==1.3.7
anthropic==0.7.2

# Database (optional, for questionnaire data)
supabase==2.0.3

# Utilities
requests==2.31.0
